Fix scoped restore route freshness
