+++ cmuxTests/MobileHostAuthorizationTests.swift
@@ -11,78 +11,6 @@ import Testing
 @Suite(.serialized)
 @MainActor
 struct MobileHostAuthorizationTests {
-    @Test func testMobileHostIdentityPrefersSharedIDAcrossBundleDefaults() throws {
-        let directory = FileManager.default.temporaryDirectory
-            .appendingPathComponent(UUID().uuidString, isDirectory: true)
-        try FileManager.default.createDirectory(at: directory, withIntermediateDirectories: true)
-        defer { try? FileManager.default.removeItem(at: directory) }
-        let sharedIDURL = directory.appendingPathComponent("mobile-host-device-id")
-        let sharedID = "3D56C547-271C-47D8-84F6-5C79C9394A37"
-        try sharedID.lowercased().write(to: sharedIDURL, atomically: true, encoding: .utf8)
-
-        let suiteName = "mobile-host-identity-\(UUID().uuidString)"
-        let defaults = try #require(UserDefaults(suiteName: suiteName))
-        defer { defaults.removePersistentDomain(forName: suiteName) }
-        defaults.set("175dff61-cabe-4076-b5ac-f5c1c04b62fa", forKey: "mobileHost.deviceID")
-
-        #expect(MobileHostIdentity.deviceID(defaults: defaults, sharedIDURL: sharedIDURL) == sharedID)
-        #expect(defaults.string(forKey: "mobileHost.deviceID") == sharedID)
-    }
-
-    @Test func testMobileHostIdentityMigratesExistingBundleIDToSharedFile() throws {
-        let directory = FileManager.default.temporaryDirectory
-            .appendingPathComponent(UUID().uuidString, isDirectory: true)
-        try FileManager.default.createDirectory(at: directory, withIntermediateDirectories: true)
-        defer { try? FileManager.default.removeItem(at: directory) }
-        let sharedIDURL = directory.appendingPathComponent("mobile-host-device-id")
-
-        let suiteName = "mobile-host-identity-\(UUID().uuidString)"
-        let defaults = try #require(UserDefaults(suiteName: suiteName))
-        defer { defaults.removePersistentDomain(forName: suiteName) }
-        let defaultID = "C2FD4C2D-E0AF-447D-A8A4-D37BF67751EF"
-        defaults.set(defaultID.lowercased(), forKey: "mobileHost.deviceID")
-
-        #expect(MobileHostIdentity.deviceID(defaults: defaults, sharedIDURL: sharedIDURL) == defaultID)
-        let persisted = try String(contentsOf: sharedIDURL, encoding: .utf8)
-        #expect(persisted == defaultID)
-    }
-
-    @Test func testMobileHostIdentityReadsExistingSharedIDWithoutDefaults() throws {
-        let directory = FileManager.default.temporaryDirectory
-            .appendingPathComponent(UUID().uuidString, isDirectory: true)
-        try FileManager.default.createDirectory(at: directory, withIntermediateDirectories: true)
-        defer { try? FileManager.default.removeItem(at: directory) }
-        let sharedIDURL = directory.appendingPathComponent("mobile-host-device-id")
-        let sharedID = "4BF9566D-5D67-4C79-8974-B42D5CF39DE9"
-        try sharedID.write(to: sharedIDURL, atomically: true, encoding: .utf8)
-
-        let suiteName = "mobile-host-identity-\(UUID().uuidString)"
-        let defaults = try #require(UserDefaults(suiteName: suiteName))
-        defer { defaults.removePersistentDomain(forName: suiteName) }
-
-        #expect(MobileHostIdentity.deviceID(defaults: defaults, sharedIDURL: sharedIDURL) == sharedID)
-        #expect(defaults.string(forKey: "mobileHost.deviceID") == sharedID)
-    }
-
-    @Test func testMobileHostIdentityRepairsInvalidSharedIDFile() throws {
-        let directory = FileManager.default.temporaryDirectory
-            .appendingPathComponent(UUID().uuidString, isDirectory: true)
-        try FileManager.default.createDirectory(at: directory, withIntermediateDirectories: true)
-        defer { try? FileManager.default.removeItem(at: directory) }
-        let sharedIDURL = directory.appendingPathComponent("mobile-host-device-id")
-        try "not-a-uuid".write(to: sharedIDURL, atomically: true, encoding: .utf8)
-
-        let suiteName = "mobile-host-identity-\(UUID().uuidString)"
-        let defaults = try #require(UserDefaults(suiteName: suiteName))
-        defer { defaults.removePersistentDomain(forName: suiteName) }
-        let fallbackID = "08E3578B-195D-486F-B874-023CDA2B647D"
-        defaults.set(fallbackID, forKey: "mobileHost.deviceID")
-
-        #expect(MobileHostIdentity.deviceID(defaults: defaults, sharedIDURL: sharedIDURL) == fallbackID)
-        #expect(defaults.string(forKey: "mobileHost.deviceID") == fallbackID)
-        #expect(try String(contentsOf: sharedIDURL, encoding: .utf8) == fallbackID)
-    }
-
     @Test func testAttachTicketStoreKeepsMultipleTicketsForSameTerminal() throws {
         let store = MobileAttachTicketStore()
         let route = try CmxAttachRoute(
